backend/auth-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
pydantic==2.10.6
pydantic-settings==2.8.1
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
pytest==7.4.2
httpx==0.24.1
bcrypt==4.0.1
python-dotenv==1.0.0
python-jose==3.4.0
email-validator==2.2.0
